Refactoring card locations and fixing js error when card is first element
