--- app/components/editor-test-hr.js
@@ -1,3 +1,0 @@
-import Component from '@ember/component';
-
-export default Component.extend({});
