--- app/templates/components/editor-test-hr-editor.hbs
@@ -1,1 +1,0 @@
-{{yield}}
