--- app/templates/components/editor-test-hr.hbs
@@ -1,1 +1,0 @@
-{{yield}}
