+++ app/templates/components/tools/text-editor.hbs
@@ -1,13 +1,9 @@
-{{log cards}}
 {{#mobiledoc-editor mobiledoc=content
                     autofocus=true
                     spellcheck=true
-                    cards=cards
                     on-change=(action "mobileDocUpdated")
                     class="note-editor d-flex flex-column-reverse"
 as |editor|}}
-  {{mobiledoc-toolbar editor=editor}}
-  {{!--
   <div class="toolbar d-flex justify-content-between p-1">
     <div class="col-4">
       <div class="btn-group">
@@ -37,5 +33,5 @@ as |editor|}}
       </div>
     </div>
   </div>
-  --}}
 {{/mobiledoc-editor}}
+
