+++ lib/koenig-editor/addon/components/koenig-card/bill.js
@@ -1,5 +1,5 @@
 import Component from '@ember/component';
-import layout from '../templates/components/koenig-card-bill';
+import layout from '../../templates/components/koenig-card/bill';
 import { computed } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { htmlSafe } from '@ember/string';
