+++ lib/koenig-editor/addon/components/koenig-card/embedly.js
@@ -1,5 +1,7 @@
 import Component from '@ember/component';
-import layout from '../templates/components/koenig-card-html';
+import layout from '../../templates/components/koenig-card/embedly';
+import { computed } from '@ember/object';
+import { htmlSafe } from '@ember/string';
 import { isBlank } from '@ember/utils';
 import { run } from '@ember/runloop';
 import { set } from '@ember/object';
@@ -19,22 +21,18 @@ export default Component.extend({
   init() {
     this._super(...arguments);
 
-    if (!this.get('payload.html')) {
-      this.set('payload.html', '');
+    if (!this.get('payload.url')) {
+      this.set('payload.url', '');
     }
   },
 
   actions: {
-    updateHtml(html) {
-      this._updatePayloadAttr('html', html);
-    },
-
-    updateCaption(caption) {
-      this._updatePayloadAttr('caption', caption);
+    updateUrl(url) {
+      this._updatePayloadAttr('url', url);
     },
 
     leaveEditMode() {
-      if (isBlank(this.get('payload.html'))) {
+      if (isBlank(this.get('payload.url'))) {
         // afterRender is required to avoid double modification of `isSelected`
         // TODO: see if there's a way to avoid afterRender
         run.scheduleOnce('afterRender', this, function() {
