+++ lib/koenig-editor/addon/components/koenig-card/hr.js
@@ -1,5 +1,5 @@
 import Component from '@ember/component';
-import layout from '../templates/components/koenig-card-hr';
+import layout from '../../templates/components/koenig-card/hr';
 
 export default Component.extend({
   tagName: '',
