+++ lib/koenig-editor/addon/components/koenig-card/html.js
@@ -1,7 +1,5 @@
 import Component from '@ember/component';
-import layout from '../templates/components/koenig-card-embedly';
-import { computed } from '@ember/object';
-import { htmlSafe } from '@ember/string';
+import layout from '../../templates/components/koenig-card/html';
 import { isBlank } from '@ember/utils';
 import { run } from '@ember/runloop';
 import { set } from '@ember/object';
@@ -21,18 +19,22 @@ export default Component.extend({
   init() {
     this._super(...arguments);
 
-    if (!this.get('payload.url')) {
-      this.set('payload.url', '');
+    if (!this.get('payload.html')) {
+      this.set('payload.html', '');
     }
   },
 
   actions: {
-    updateUrl(url) {
-      this._updatePayloadAttr('url', url);
+    updateHtml(html) {
+      this._updatePayloadAttr('html', html);
+    },
+
+    updateCaption(caption) {
+      this._updatePayloadAttr('caption', caption);
     },
 
     leaveEditMode() {
-      if (isBlank(this.get('payload.url'))) {
+      if (isBlank(this.get('payload.html'))) {
         // afterRender is required to avoid double modification of `isSelected`
         // TODO: see if there's a way to avoid afterRender
         run.scheduleOnce('afterRender', this, function() {
