+++ lib/koenig-editor/addon/components/koenig-card/image.js
@@ -1,6 +1,6 @@
 import $ from 'jquery';
 import Component from '@ember/component';
-import layout from '../templates/components/koenig-card-image';
+import layout from '../../templates/components/koenig-card/image';
 import {
   IMAGE_EXTENSIONS,
   IMAGE_MIME_TYPES
