+++ lib/koenig-editor/addon/components/koenig-card/markdown.js
@@ -1,6 +1,6 @@
 import Component from '@ember/component';
 import formatMarkdown from 'ghost-admin/utils/format-markdown';
-import layout from '../templates/components/koenig-card-markdown';
+import layout from '../../templates/components/koenig-card/markdown';
 import { computed } from '@ember/object';
 import { htmlSafe } from '@ember/string';
 import { isBlank } from '@ember/utils';
