+++ lib/koenig-editor/addon/components/koenig-editor.js
@@ -38,9 +38,9 @@ export const BLANK_DOC = {
 
 // map card names to component names
 export const CARD_COMPONENT_MAP = {
-  hr: 'koenig-card-hr',
-  embedly: 'koenig-card-embedly',
-  bill: 'koenig-card-bill'
+  hr: 'koenig-card/hr',
+  embedly: 'koenig-card/embedly',
+  bill: 'koenig-card/bill'
 };
 
 const CURSOR_BEFORE = -1;
@@ -733,12 +733,14 @@ export default Component.extend({
       if (cursorDirection === CURSOR_BEFORE) {
         nextPosition = section.prev.tailPosition();
       } else {
-        nextPosition = section.next.headPosition();
+        if (section.next) {
+          nextPosition = section.next.headPosition();
+        }
       }
 
       postEditor.removeSection(section);
 
-      if (cursorDirection !== NO_CURSOR_MOVEMENT) {
+      if (cursorDirection !== NO_CURSOR_MOVEMENT && nextPosition) {
         postEditor.setRange(nextPosition);
       }
     });
