--- lib/koenig-editor/addon/templates/components/koenig-card-embedly.hbs
@@ -1,15 +1,0 @@
-{{#koenig-card
-    tagName="div"
-    class=(concat (kg-style "media-card-h") " flex flex-column")
-    isSelected=isSelected
-    isEditing=isEditing
-    selectCard=(action selectCard)
-    editCard=(action editCard)
-    onLeaveEdit=(action "leaveEditMode")
-}}
-  {{#if isEditing}}
-    {{input class="form-control" value=payload.url update=(action "updateUrl")}}
-  {{else}}
-    <a data-card-controls="0" href={{payload.url}} class="embedly-card">Embedly</a>
-  {{/if}}
-{{/koenig-card}}
