+++ lib/koenig-editor/addon/templates/components/koenig-card/hr.hbs
