+++ lib/koenig-editor/addon/templates/components/koenig-card/html.hbs
