+++ lib/koenig-editor/addon/templates/components/koenig-card/image.hbs
