+++ lib/koenig-editor/addon/templates/components/koenig-card/markdown.hbs
