+++ lib/koenig-editor/addon/templates/components/koenig-toolbar.hbs
@@ -62,14 +62,3 @@
     {{fa-icon "quote-right"}}
   </button>
 </div>
-<div class="btn-group mr-2" role="group" aria-label="Fourth group">
-
-  <button
-    type="button"
-    title="Link"
-    class="btn btn-secondary koenig-toolbar-btn koenig-toolbar-btn-link"
-  >
-    {{fa-icon "link"}}
-  </button>
-
-</div>
