--- lib/koenig-editor/app/components/koenig-card-hr.js
@@ -1,1 +1,0 @@
-export { default } from 'koenig-editor/components/koenig-card-hr';
