--- lib/koenig-editor/app/components/koenig-card-image.js
@@ -1,1 +1,0 @@
-export { default } from 'koenig-editor/components/koenig-card-image';
