+++ lib/koenig-editor/app/components/koenig-card/markdown.js
@@ -1,0 +1,1 @@
+export { default } from 'koenig-editor/components/koenig-card/markdown';
